# backend/docs/swagger.yaml
basePath: /
definitions:
  models.AddPackResponse:
    properties:
      author:
        example: author
        type: string
      creation_date:
        example: creation_date
        type: string
      name:
        example: name
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: error
        type: string
      message:
        example: message
        type: string
    type: object
  models.PackListResponse:
    properties:
      packs:
        items:
          $ref: '#/definitions/models.PackResponse'
        type: array
      packs_num:
        example: 0
        type: integer
    type: object
  models.PackResponse:
    properties:
      author:
        example: author
        type: string
      creation_date:
        example: creation_date
        type: string
      downloads_num:
        example: 0
        type: integer
      file_size:
        example: 0
        type: integer
      guid:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      name:
        example: name
        type: string
      tags:
        example:
        - tag1
        - tag2
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: almiluk@gmail.com
    name: almiluk
  title: API for managing question packs for the 'SIGame' game
  version: 0.0.0
paths:
  /api/v1/packs:
    get:
      consumes:
      - application/json
      description: List packs with filters
      parameters:
      - example: author
        in: query
        name: author
        type: string
      - example: 01.01.1970
        in: query
        name: max_creation_date
        type: string
      - example: 01.01.1970
        in: query
        name: min_creation_date
        type: string
      - example: name
        in: query
        name: name
        type: string
      - enum:
        - creation_date
        - downloads_num
        example: creation_date
        in: query
        name: sort_by
        type: string
      - example: tag1,tag2
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PackListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List packs
      tags:
      - packs
    post:
      consumes:
      - multipart/form-data
      description: Add new questions pack
      parameters:
      - description: Pack data
        in: formData
        name: pack
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddPackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.PackResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add pack
      tags:
      - packs
  /packs/{guid}:
    get:
      description: Download questions pack
      parameters:
      - description: Pack guid
        format: uuid
        in: path
        name: guid
        required: true
        type: string
      - description: Wanted package file name
        in: query
        name: filename
        type: string
      produces:
      - application/octet-stream
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Download pack
      tags:
      - packs
swagger: "2.0"